feat: crée un panel admin externe et supprime l'ancien panel bugué
