+++ admin-app/next.config.js
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  images: {
+    domains: ['localhost', 'vercel.app'],
+  },
+}
+
+module.exports = nextConfig
