--- app/admin/components/ErrorBoundary.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import React from 'react';
-
-const ErrorBoundary: React.FC<{children: React.ReactNode}> = ({ children }) => {
-  return <>{children}</>;
-};
-
-export default ErrorBoundary;
